feat: 新增7个功能模块 + 修复测试
新增模块:
- accessaudit: 访问审计系统
- aiassistant: AI智能助手
- analytics: 数据分析平台
- compliance: 合规检查引擎
- dedupviz: 去重可视化
- filesnapmgr: 文件快照管理
- permissioncenter: 权限中心
- sysaudit: 系统审计

修复:
- assetmgr: GetAssetSummary groups计数
- cloudsync/clustermanager/containersec/costgovernance/migration/smarthdd: 测试修复
- filecache: 新增缓存管理器
- taskqueue: DAG任务编排 + handler

+++ internal/costgovernance/costgovernance_test.go
@@ -230,7 +230,7 @@ func TestAnalyzerPredictCost(t *testing.T) {
 	require.NoError(t, err)
 	assert.Len(t, predictions, 7)
 	// 趋势上升，预测值应大于最后数据点
-	assert.Greater(t, predictions[6].Cost, 1000+9*50)
+	assert.Greater(t, predictions[6].Cost, float64(1000+9*50))
 }
 
 func TestAnalyzerDetectAnomalies(t *testing.T) {
